Refactored bomberman and trading card examples
